Finalize essential unit tests
