+++ tests/__main__.py
@@ -1,0 +1,8 @@
+import unittest
+from .test_file_creation import TestFileCreation
+from .test_log_creation import TestLogCreation
+from .test_file_name import TestFileName
+from .test_file_hash import TestFileHash
+
+if __name__ == "__main__":
+    unittest.main()
